pyproject.toml
[project]
name = "shu-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi-cache2[redis]>=0.2.2",
    # "bcrypt==4.0.1",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
]

[dependency-groups]
dev = ["ruff>=0.11.5", "ipykernel>=6.29.5"]
